Switching from i64 to i128 due to overflow errors.

+++ src/util.rs
@@ -10,11 +10,11 @@ use crate::constants::*;
 
 // generates an element of Z_q[X]/(X^d+1)
 // TODO: more sophisticated / cleaner random polynomial sampling
-pub fn generate_polynomial(q : i64, d : i64) -> Polynomial<i64> {
+pub fn generate_polynomial(q : i128, d : i128) -> Polynomial<i128> {
     let max_degree = d-1; // TODO is this REALLY the max degree? I think NOT. Fix later.
     let mut rng = rand::thread_rng();
     //let num_terms = rng.gen_range(1, max_degree + 2);
-    let mut coefficient_vec: Vec<i64> = Vec::new();
+    let mut coefficient_vec: Vec<i128> = Vec::new();
     for degree in 0..max_degree {
         coefficient_vec.push(rng.gen_range(0..q));
     }
@@ -23,7 +23,7 @@ pub fn generate_polynomial(q : i64, d : i64) -> Polynomial<i64> {
 }
 
 
-pub fn generate_random_matrix(rows : i64, cols : i64, q : i64, d : i64) -> Array2<Polynomial<i64>> {
+pub fn generate_random_matrix(rows : i128, cols : i128, q : i128, d : i128) -> Array2<Polynomial<i128>> {
     let mut mat = Array2::from_elem((rows as usize,cols as usize), Polynomial::new(vec![])); 
     for i in 0..rows {
         for j in 0..cols {
@@ -40,17 +40,17 @@ pub fn generate_random_matrix(rows : i64, cols : i64, q : i64, d : i64) -> Array
 // TODO yes we currently have to clone coeff_dist in when we want to re-use it, but it's NBD
 // because it's usually very small. We don't want to use a mutable ref because that would be weird,
 // since we don't want to change the underlying distribution data / remove elements if we're gonna re-use it later...
-pub fn generate_polynomial_picky(q : i64, d : usize, mut coeff_dist : Vec<i64>) -> Polynomial<i64> {
+pub fn generate_polynomial_picky(q : i128, d : usize, mut coeff_dist : Vec<i128>) -> Polynomial<i128> {
     assert!(coeff_dist.len() == d, "Must have one coefficient for each degree of polynomial");
     let mut rng = rand::thread_rng();
-    let mut coefficient_vec: Vec<i64> = Vec::new();
+    let mut coefficient_vec: Vec<i128> = Vec::new();
     for degree in 0..d {
         let random_index : usize = *(coeff_dist.as_slice().choose(&mut rng).unwrap()) as usize;
         let coeff = coeff_dist[random_index];
         coeff_dist.remove(random_index);
         let signed : bool = rng.gen();
         if (coeff > 0) && signed {
-            coefficient_vec.push((-1*coeff) as i64);
+            coefficient_vec.push((-1*coeff) as i128);
         }
         else {
             coefficient_vec.push(coeff);
@@ -62,7 +62,7 @@ pub fn generate_polynomial_picky(q : i64, d : usize, mut coeff_dist : Vec<i64>) 
 
 
 // Conjugation automorphism applied to a vector of polynomials
-pub fn sigma_inv_vec(vec: &Vec<Polynomial<i64>>) -> Vec<Polynomial<i64>> {
+pub fn sigma_inv_vec(vec: &Vec<Polynomial<i128>>) -> Vec<Polynomial<i128>> {
     let mut res = vec![];
     for i in 0..vec.len() {
         res.push(sigma_inv(&vec[i]));
@@ -75,11 +75,11 @@ pub fn sigma_inv_vec(vec: &Vec<Polynomial<i64>>) -> Vec<Polynomial<i64>> {
 // The "Conjugation automorphism"
 // This basically replaces all powers of X^n with X^{-n} and then reduces modulo X^d+1 (R's
 // modulus) 
-pub fn sigma_inv(a : &Polynomial<i64>) -> Polynomial<i64> {
+pub fn sigma_inv(a : &Polynomial<i128>) -> Polynomial<i128> {
 
-    let poly_data_vec : Vec<i64> = a.data().to_vec();
+    let poly_data_vec : Vec<i128> = a.data().to_vec();
 
-    let mut new_coeff_vec : Vec<i64> = vec![0; 64]; 
+    let mut new_coeff_vec : Vec<i128> = vec![0; 64]; 
 
     // terms of degree 0 unchanged
     // degree 1-64, we simply transform X^n to -X^{64-n}.
@@ -100,8 +100,8 @@ pub fn sigma_inv(a : &Polynomial<i64>) -> Polynomial<i64> {
 
 
 
-pub fn multiply_poly_vec_ints(p_vec : &Vec<Polynomial<i64>>, ints: &Vec<i64>) -> Vec<Polynomial<i64>> {
-    let mut p_vec_res : Vec<Polynomial<i64>> = vec![];
+pub fn multiply_poly_vec_ints(p_vec : &Vec<Polynomial<i128>>, ints: &Vec<i128>) -> Vec<Polynomial<i128>> {
+    let mut p_vec_res : Vec<Polynomial<i128>> = vec![];
 
     for p in p_vec {
         p_vec_res.push(multiply_poly_ints(p, ints));
@@ -109,8 +109,8 @@ pub fn multiply_poly_vec_ints(p_vec : &Vec<Polynomial<i64>>, ints: &Vec<i64>) ->
     p_vec_res
 }
 
-pub fn multiply_poly_ints(p : &Polynomial<i64>, ints: &Vec<i64>) -> Polynomial<i64> {
-    let mut p_res : Polynomial<i64> = Polynomial::new(vec![]);
+pub fn multiply_poly_ints(p : &Polynomial<i128>, ints: &Vec<i128>) -> Polynomial<i128> {
+    let mut p_res : Polynomial<i128> = Polynomial::new(vec![]);
 
     for coeff in ints {
         p_res = p_res + scale_polynomial(p, *coeff as f32);
@@ -120,17 +120,17 @@ pub fn multiply_poly_ints(p : &Polynomial<i64>, ints: &Vec<i64>) -> Polynomial<i
 
 
 // randomly samples n integers mod q, returns them as a vec
-pub fn random_sample_Z_q(n : i64, q: i64) -> Vec<i64> {
+pub fn random_sample_Z_q(n : i128, q: i128) -> Vec<i128> {
 
     let mut rng = thread_rng();
     let dist = Uniform::from(0..q);
 
-    let sample: Vec<i64> = (0..n).map(|_| rng.sample(dist)).collect();
+    let sample: Vec<i128> = (0..n).map(|_| rng.sample(dist)).collect();
     sample
 }
 
 // scales a given polynomial by a scale factor (usually < 1)
-pub fn scale_poly_vec(vec : &Vec<Polynomial<i64>>, s : f32) -> Vec<Polynomial<i64>> {
+pub fn scale_poly_vec(vec : &Vec<Polynomial<i128>>, s : f32) -> Vec<Polynomial<i128>> {
     let mut new_vec = vec![];
     for i in 0..vec.len() {
         new_vec.push(scale_polynomial(&vec[i], s));
@@ -140,14 +140,14 @@ pub fn scale_poly_vec(vec : &Vec<Polynomial<i64>>, s : f32) -> Vec<Polynomial<i6
 
 
 // scales a given polynomial by a scale factor (usually < 1)
-pub fn scale_polynomial(p : &Polynomial<i64>, s : f32) -> Polynomial<i64> {
+pub fn scale_polynomial(p : &Polynomial<i128>, s : f32) -> Polynomial<i128> {
     //let constant_polynomial : Polynomial<f32> = Polynomial::new(vec![s]);
-    let poly_vec : Vec<i64> = p.data().to_vec(); 
-    let scaled_poly_vec : Vec<i64> = poly_vec.iter().map(|&x| (((x as f32) * s).floor() as i64)).collect();
+    let poly_vec : Vec<i128> = p.data().to_vec(); 
+    let scaled_poly_vec : Vec<i128> = poly_vec.iter().map(|&x| (((x as f32) * s).floor() as i128)).collect();
     return Polynomial::new(scaled_poly_vec);
 }
 
-pub fn vec_poly_norm_squared(vec: &Vec<Polynomial<i64>>) -> f64 {
+pub fn vec_poly_norm_squared(vec: &Vec<Polynomial<i128>>) -> f64 {
     vec.iter()
         .map(|poly| poly_norm(poly))  // Compute the squared norm of each polynomial
         .sum()  // Sum the squared norms
@@ -155,27 +155,27 @@ pub fn vec_poly_norm_squared(vec: &Vec<Polynomial<i64>>) -> f64 {
 
 
 // takes the 2-norm of a given polynomial (squared)
-pub fn poly_norm(p : &Polynomial<i64>) -> f64 {
-    let norm : i64 = p.data().to_vec().iter().map(|&x| x*x).sum::<i64>();
+pub fn poly_norm(p : &Polynomial<i128>) -> f64 {
+    let norm : i128 = p.data().to_vec().iter().map(|&x| x*x).sum::<i128>();
     norm as f64
 }
 
 // Does NOT square the norm
-pub fn poly_norm_strict(p : &Polynomial<i64>) -> f64 {
-    let strict_norm : f64 = ((p.data().to_vec().iter().map(|&x| x*x).sum::<i64>()) as f64).sqrt();
+pub fn poly_norm_strict(p : &Polynomial<i128>) -> f64 {
+    let strict_norm : f64 = ((p.data().to_vec().iter().map(|&x| x*x).sum::<i128>()) as f64).sqrt();
     strict_norm
 }
 
 
 // Does NOT square the norm. Integer vec version
-pub fn l2_norm(vec : &Vec<i64>) -> f64 {
-    let sum_of_squares: i64 = vec.iter().map(|&x| x * x).sum();
+pub fn l2_norm(vec : &Vec<i128>) -> f64 {
+    let sum_of_squares: i128 = vec.iter().map(|&x| x * x).sum();
     (sum_of_squares as f64).sqrt()
 }
 
 // TODO we're using a statistical estimate of the sup to accomplish this, not sure if that's valid.
 // should be since there's basically no other way to feasibly do this (I think). Check this.
-pub fn operator_norm(c : &Polynomial<i64>) -> f64 {
+pub fn operator_norm(c : &Polynomial<i128>) -> f64 {
     // here's the play.
     // compute n_samples sample polynomials r (where n_samples is some large number, but not too large)
     // compute the described ratio / inner product
@@ -196,7 +196,7 @@ pub fn operator_norm(c : &Polynomial<i64>) -> f64 {
     sup_estimate
 }
 
-pub fn compute_total_norm(projection: &Array2<Polynomial<i64>>) -> f64 {
+pub fn compute_total_norm(projection: &Array2<Polynomial<i128>>) -> f64 {
     let mut total_norm_squared: f64 = 0.0;
     for row in projection.outer_iter() {
         for poly in row.iter() {
@@ -207,9 +207,9 @@ pub fn compute_total_norm(projection: &Array2<Polynomial<i64>>) -> f64 {
     return f64::sqrt(total_norm_squared);
 }
 
-pub fn vec_inner_product(v1: &Vec<i64>, v2: &Vec<i64>) -> i64 {
+pub fn vec_inner_product(v1: &Vec<i128>, v2: &Vec<i128>) -> i128 {
     assert!(v1.len() == v2.len(), "inner product not defined on vectors of unequal length");
-    let mut result : i64 = 0;
+    let mut result : i128 = 0;
     for i in 0..v1.len() {
         let product = &v1[i] * &v2[i];
         result += product;
@@ -220,8 +220,8 @@ pub fn vec_inner_product(v1: &Vec<i64>, v2: &Vec<i64>) -> i64 {
 
 // elementwise multiplication of a vector of polynomials by a single polynomial. NOT an inner
 // product.
-pub fn poly_by_poly_vec(poly: &Polynomial<i64>, v1: &Vec<Polynomial<i64>>) -> Vec<Polynomial<i64>> {
-    let mut v2 : Vec<Polynomial<i64>> = vec![];
+pub fn poly_by_poly_vec(poly: &Polynomial<i128>, v1: &Vec<Polynomial<i128>>) -> Vec<Polynomial<i128>> {
+    let mut v2 : Vec<Polynomial<i128>> = vec![];
 
     for i in 0..v1.len() {
         v2.push(poly * &v1[i]);
@@ -230,9 +230,9 @@ pub fn poly_by_poly_vec(poly: &Polynomial<i64>, v1: &Vec<Polynomial<i64>>) -> Ve
 }
 
 
-pub fn add_poly_vec(v1: &Vec<Polynomial<i64>>, v2: &Vec<Polynomial<i64>>) -> Vec<Polynomial<i64>> {
+pub fn add_poly_vec(v1: &Vec<Polynomial<i128>>, v2: &Vec<Polynomial<i128>>) -> Vec<Polynomial<i128>> {
     assert!(v1.len() == v2.len(), "summation not defined on vectors of unequal length");
-    let mut v_res : Vec<Polynomial<i64>> = vec![];
+    let mut v_res : Vec<Polynomial<i128>> = vec![];
     for i in 0..v1.len() {
         let new_p = &v1[i] + &v2[i];
         v_res.push(new_p);
@@ -240,9 +240,9 @@ pub fn add_poly_vec(v1: &Vec<Polynomial<i64>>, v2: &Vec<Polynomial<i64>>) -> Vec
     v_res
 }
 
-pub fn add_vecs(v1: &Vec<i64>, v2: &Vec<i64>) -> Vec<i64> {
+pub fn add_vecs(v1: &Vec<i128>, v2: &Vec<i128>) -> Vec<i128> {
     assert!(v1.len() == v2.len(), "summation not defined on vectors of unequal length");
-    let mut v_res : Vec<i64> = vec![];
+    let mut v_res : Vec<i128> = vec![];
     for i in 0..v1.len() {
         let new_p = &v1[i] + &v2[i];
         v_res.push(new_p);
@@ -254,8 +254,8 @@ pub fn add_vecs(v1: &Vec<i64>, v2: &Vec<i64>) -> Vec<i64> {
 
 /* Rarely, we need to do elementwise addition of a single polynomial to an entire vector of
  * polynomials.. kind of strange, but how this is described in the protocol */
-pub fn add_poly_vec_by_poly(v1: &Vec<Polynomial<i64>>, p: &Polynomial<i64>) -> Vec<Polynomial<i64>> {
-    let mut v_res : Vec<Polynomial<i64>> = vec![];
+pub fn add_poly_vec_by_poly(v1: &Vec<Polynomial<i128>>, p: &Polynomial<i128>) -> Vec<Polynomial<i128>> {
+    let mut v_res : Vec<Polynomial<i128>> = vec![];
     for i in 0..v1.len() {
         let new_p = &v1[i] + p;
         v_res.push(new_p);
@@ -267,8 +267,8 @@ pub fn add_poly_vec_by_poly(v1: &Vec<Polynomial<i64>>, p: &Polynomial<i64>) -> V
 
 
 
-pub fn gen_empty_poly_vec(n : usize) -> Vec<Polynomial<i64>> {
-    let mut v_res : Vec<Polynomial<i64>> = vec![];
+pub fn gen_empty_poly_vec(n : usize) -> Vec<Polynomial<i128>> {
+    let mut v_res : Vec<Polynomial<i128>> = vec![];
     for i in 0..n {
         v_res.push(Polynomial::new(vec![]));
     }
@@ -277,7 +277,7 @@ pub fn gen_empty_poly_vec(n : usize) -> Vec<Polynomial<i64>> {
 }
 
 // used specifically do decompose e.g., the entirety of t_i = t_i^(0) + ... + t_i^(t_1-1)b_1^(t_1-1)
-pub fn decompose_polynomial_vec(vec : &Vec<Polynomial<i64>>, base : i64, exp: i64) -> Vec<Vec<Polynomial<i64>>> {
+pub fn decompose_polynomial_vec(vec : &Vec<Polynomial<i128>>, base : i128, exp: i128) -> Vec<Vec<Polynomial<i128>>> {
 
     // TODO KAPPA is hardcoded here for now. Fix later.
     let mut res = vec![vec![Polynomial::new(vec![]); KAPPA as usize]; exp as usize];
@@ -296,7 +296,7 @@ pub fn decompose_polynomial_vec(vec : &Vec<Polynomial<i64>>, base : i64, exp: i6
 
 // computes the centered representative of a polynomial coefficient wrt a given base, i.e.,
 // returns a value in the range of [-b/2, b/2]
-pub fn centered_rep(mut val : i64, b: i64) -> i64 {
+pub fn centered_rep(mut val : i128, b: i128) -> i128 {
     if val > b / 2 {
         val -= b;
     }
@@ -307,12 +307,12 @@ pub fn centered_rep(mut val : i64, b: i64) -> i64 {
 }
 
 
-pub fn decompose_polynomial(p : &Polynomial<i64>, base : i64, exp: i64) -> Vec<Polynomial<i64>> {
+pub fn decompose_polynomial(p : &Polynomial<i128>, base : i128, exp: i128) -> Vec<Polynomial<i128>> {
 
-    let poly_data_vec : Vec<i64> = p.data().to_vec();
+    let poly_data_vec : Vec<i128> = p.data().to_vec();
 
     // this takes the form of, e.g., g_{ij} = g_{ij}^(0) + ... + g_{ij}^{t_2-1}b_2^{t_2-1}
-    let mut decomp_vec : Vec<Polynomial<i64>> = vec![Polynomial::<i64>::new(vec![]); exp as usize];
+    let mut decomp_vec : Vec<Polynomial<i128>> = vec![Polynomial::<i128>::new(vec![]); exp as usize];
 
     // we decompose each coefficient a_j of the polynomial (which we can express as a base K 
 
@@ -334,15 +334,15 @@ pub fn decompose_polynomial(p : &Polynomial<i64>, base : i64, exp: i64) -> Vec<P
 
 // Used to convert a vector of polynomials \vec{\b{s}}_i to another vector
 // which is a concatenation of all the coefficients 
-pub fn witness_coeff_concat(vec: &Vec<Polynomial<i64>>) -> Vec<i64> {
+pub fn witness_coeff_concat(vec: &Vec<Polynomial<i128>>) -> Vec<i128> {
     // NOTE: we assume that the polynomials in vec will be of degree D... 
     // But we want to return a vec of degree N*D (where N is the assumed length of vec)..
     // So we recognize that for a polynomial ring mod X^D+1, X^d congruent to -1... so we can
     // rewrite as such.
-    let mut coeffs : Vec<i64> = vec![];
+    let mut coeffs : Vec<i128> = vec![];
     for j in 0..vec.len() {
-        let poly : &Polynomial<i64> = &vec[j];
-        let poly_vec_data : Vec<i64> = poly.data().to_vec();
+        let poly : &Polynomial<i128> = &vec[j];
+        let poly_vec_data : Vec<i128> = poly.data().to_vec();
         
         // TODO can we simplify this to poly_vec_data.len()? I don't think so, since data might
         // store fewer coeffs
@@ -376,7 +376,7 @@ pub fn witness_coeff_concat(vec: &Vec<Polynomial<i64>>) -> Vec<i64> {
 // Basically does the opposite of the above function
 // Takes a vec of straight ints of dimension N*D for some int N, and returns
 // N length vec of reconstructed Polynomials using the coefficients
-pub fn concat_coeff_reduction(vec: &Vec<i64>) -> Vec<Polynomial<i64>> {
+pub fn concat_coeff_reduction(vec: &Vec<i128>) -> Vec<Polynomial<i128>> {
     assert!((vec.len() % (D as usize)) == 0, "dimension of vec is wrong");
     let n = vec.len() / (D as usize);
     let mut res = vec![];
@@ -400,7 +400,7 @@ where
 }
 
 // compute the dot product between two lists of polynomials
-pub fn polynomial_vec_inner_product(v1: &[Polynomial<i64>], v2: &[Polynomial<i64>]) -> Polynomial<i64> {
+pub fn polynomial_vec_inner_product(v1: &[Polynomial<i128>], v2: &[Polynomial<i128>]) -> Polynomial<i128> {
     assert!(v1.len() == v2.len(), "inner product not defined on vectors of unequal length");
     let mut result = Polynomial::new(vec![]);
     for i in 0..v1.len() {
@@ -410,7 +410,7 @@ pub fn polynomial_vec_inner_product(v1: &[Polynomial<i64>], v2: &[Polynomial<i64
     result
 }
 
-pub fn matmul(m1: &Array2<i64>, m2: &Array2<i64>) -> Array2<i64> {
+pub fn matmul(m1: &Array2<i128>, m2: &Array2<i128>) -> Array2<i128> {
     assert!(m1.shape()[1] == m2.shape()[0], "matrix product not defined on matrices which do not have dimension (m,n) x (n,k), for some m,n,k");
     let m = m1.shape()[0];
     let k = m2.shape()[1];
@@ -427,7 +427,7 @@ pub fn matmul(m1: &Array2<i64>, m2: &Array2<i64>) -> Array2<i64> {
 }
 
 // compute the matrix product between two polynomial MATRICES (Array2, slightly different)
-pub fn polynomial_matrix_product(m1: &Array2<Polynomial<i64>>, m2: &Array2<Polynomial<i64>>) -> Array2<Polynomial<i64>> {
+pub fn polynomial_matrix_product(m1: &Array2<Polynomial<i128>>, m2: &Array2<Polynomial<i128>>) -> Array2<Polynomial<i128>> {
     assert!(m1.shape()[1] == m2.shape()[0], "matrix product not defined on matrices which do not have dimension (m,n) x (n,k), for some m,n,k");
     let m = m1.shape()[0];
     let k = m2.shape()[1];
